requirements.txt
Django==4.0.10
djangorestframework==3.14.0
drf-spectacular==0.26.0
django-redis==5.2.0
psycopg2-binary==2.9.3
whitenoise==5.3.0
python-decouple==3.8